requirements.txt
numpy
scikit-learn
matplotlib
seaborn
spacy
nltk
hdbscan
requests
python-dotenv
PyPDF2
bertopic
umap-learn
streamlit
plotly
pandas
